backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "google-adk>=1.17.0",
    "google-api-python-client>=2.186.0",
    "google-cloud-secret-manager>=2.25.0",
    "google-genai>=1.47.0",
    "googlesearch-python>=1.3.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.20",
    "supabase>=2.23.1",
    "uvicorn[standard]>=0.38.0",
]
